requirements.txt
atlassian-python-api==3.41.2
certifi==2023.7.22
charset-normalizer==3.2.0
Deprecated==1.2.14
idna==3.4
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
urllib3==2.0.4
wrapt==1.15.0
